Gearing up for Claude to write this react hook
