+++ package.json
@@ -2,12 +2,19 @@
   "name": "mcp-remote",
   "version": "0.0.4",
   "type": "module",
-  "bin": "dist/proxy.js",
+  "bin": "dist/cli/proxy.js",
   "files": [
     "dist",
     "README.md",
     "LICENSE"
   ],
+  "exports": {
+    "./react": {
+      "types": "./dist/react/index.d.ts",
+      "require": "./dist/react/index.js",
+      "import": "./dist/react/index.js"
+    }
+  },
   "scripts": {
     "build": "tsup"
   },
@@ -20,6 +27,7 @@
     "@types/express": "^5.0.0",
     "@types/node": "^22.13.10",
     "prettier": "^3.5.3",
+    "react": "^19.0.0",
     "tsup": "^8.4.0",
     "tsx": "^4.19.3",
     "typescript": "^5.8.2"
