+++ src/cli/shared.ts
@@ -9,11 +9,8 @@ import fs from 'fs/promises'
 import path from 'path'
 import os from 'os'
 import crypto from 'crypto'
-import { EventEmitter } from 'events'
 import net from 'net'
-import { OAuthClientProvider, UnauthorizedError } from '@modelcontextprotocol/sdk/client/auth.js'
-import { SSEClientTransport } from '@modelcontextprotocol/sdk/client/sse.js'
-import { Transport } from '@modelcontextprotocol/sdk/shared/transport.js'
+import {OAuthClientProvider} from '@modelcontextprotocol/sdk/client/auth.js'
 import {
   OAuthClientInformation,
   OAuthClientInformationFull,
@@ -21,24 +18,7 @@ import {
   OAuthTokens,
   OAuthTokensSchema,
 } from '@modelcontextprotocol/sdk/shared/auth.js'
-
-/**
- * Options for creating an OAuth client provider
- */
-export interface OAuthProviderOptions {
-  /** Server URL to connect to */
-  serverUrl: string
-  /** Port for the OAuth callback server */
-  callbackPort: number
-  /** Path for the OAuth callback endpoint */
-  callbackPath?: string
-  /** Directory to store OAuth credentials */
-  configDir?: string
-  /** Client name to use for OAuth registration */
-  clientName?: string
-  /** Client URI to use for OAuth registration */
-  clientUri?: string
-}
+import {OAuthCallbackServerOptions, OAuthProviderOptions} from "../lib/types.js";
 
 /**
  * Implements the OAuthClientProvider interface for Node.js environments.
@@ -226,18 +206,6 @@ export class NodeOAuthClientProvider implements OAuthClientProvider {
 }
 
 /**
- * OAuth callback server setup options
- */
-export interface OAuthCallbackServerOptions {
-  /** Port for the callback server */
-  port: number
-  /** Path for the callback endpoint */
-  path: string
-  /** Event emitter to signal when auth code is received */
-  events: EventEmitter
-}
-
-/**
  * Sets up an Express server to handle OAuth callbacks
  * @param options The server options
  * @returns An object with the server, authCode, and waitForAuthCode function
@@ -282,100 +250,6 @@ export function setupOAuthCallbackServer(options: OAuthCallbackServerOptions) {
   }
 
   return { server, authCode, waitForAuthCode }
-}
-
-/**
- * Creates and connects to a remote SSE server with OAuth authentication
- * @param serverUrl The URL of the remote server
- * @param authProvider The OAuth client provider
- * @param waitForAuthCode Function to wait for the auth code
- * @returns The connected SSE client transport
- */
-export async function connectToRemoteServer(
-  serverUrl: string,
-  authProvider: OAuthClientProvider,
-  waitForAuthCode: () => Promise<string>,
-): Promise<SSEClientTransport> {
-  console.error('Connecting to remote server:', serverUrl)
-  const url = new URL(serverUrl)
-  const transport = new SSEClientTransport(url, { authProvider })
-
-  try {
-    await transport.start()
-    console.error('Connected to remote server')
-    return transport
-  } catch (error) {
-    if (error instanceof UnauthorizedError || (error instanceof Error && error.message.includes('Unauthorized'))) {
-      console.error('Authentication required. Waiting for authorization...')
-
-      // Wait for the authorization code from the callback
-      const code = await waitForAuthCode()
-
-      try {
-        console.error('Completing authorization...')
-        await transport.finishAuth(code)
-
-        // Create a new transport after auth
-        const newTransport = new SSEClientTransport(url, { authProvider })
-        await newTransport.start()
-        console.error('Connected to remote server after authentication')
-        return newTransport
-      } catch (authError) {
-        console.error('Authorization error:', authError)
-        throw authError
-      }
-    } else {
-      console.error('Connection error:', error)
-      throw error
-    }
-  }
-}
-
-/**
- * Creates a bidirectional proxy between two transports
- * @param params The transport connections to proxy between
- */
-export function mcpProxy({ transportToClient, transportToServer }: { transportToClient: Transport; transportToServer: Transport }) {
-  let transportToClientClosed = false
-  let transportToServerClosed = false
-
-  transportToClient.onmessage = (message) => {
-    console.error('[Local→Remote]', message.method || message.id)
-    transportToServer.send(message).catch(onServerError)
-  }
-
-  transportToServer.onmessage = (message) => {
-    console.error('[Remote→Local]', message.method || message.id)
-    transportToClient.send(message).catch(onClientError)
-  }
-
-  transportToClient.onclose = () => {
-    if (transportToServerClosed) {
-      return
-    }
-
-    transportToClientClosed = true
-    transportToServer.close().catch(onServerError)
-  }
-
-  transportToServer.onclose = () => {
-    if (transportToClientClosed) {
-      return
-    }
-    transportToServerClosed = true
-    transportToClient.close().catch(onClientError)
-  }
-
-  transportToClient.onerror = onClientError
-  transportToServer.onerror = onServerError
-
-  function onClientError(error: Error) {
-    console.error('Error from local client:', error)
-  }
-
-  function onServerError(error: Error) {
-    console.error('Error from remote server:', error)
-  }
 }
 
 /**
