+++ tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  entry: ['src/client.ts', 'src/proxy.ts'],
+  entry: ['src/cli/client.ts', 'src/cli/proxy.ts', 'src/react/index.ts'],
   format: ['esm'],
   dts: true,
   clean: true,
